requirements.txt
WTForms~=3.0.1
PyJWT~=2.3.0
Flask~=2.0.2